[2024.01.26] - pay_service 수정 - 신규코드 반영

+++ lib/ui/signal/signal_top_m_page.dart
@@ -808,7 +808,7 @@ class SignalMTopPageState extends State<SignalMTopPage> {
       });
     } else if (topPrf) {
       _vList2.sort((a, b) {
-        return int.parse(b.winningRate).compareTo(int.parse(a.winningRate));
+        return double.parse(b.winningRate).compareTo(double.parse(a.winningRate));
       });
     } else if (topSta) {
       _vList3.sort((a, b) {
